backend/requirements.txt
fastapi
uvicorn
python-pptx
python-multipart
pydantic
google-generativeai
python-dotenv
google-genai